+++ src/features/api/graphql/queries/GetPostsQuery.ts
