+++ src/features/api/graphql/queries/GetRecentPostsQuery.ts
